requirements.txt
# API Serving & General
fastapi
uvicorn[standard]
sqlalchemy
oracledb
tqdm

# Data Handling & NLP
pandas
numpy
fuzzywuzzy
python-Levenshtein

# Model Training & Evaluation
torch
transformers
datasets
optuna
seqeval
GitPython